requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
cryptography==3.4.8
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.62.0
h11==0.12.0
idna==3.2
jwt==1.2.0
motor==2.5.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
pymongo==3.12.0
python-decouple==3.3
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
starlette==0.13.6
typing-extensions==3.10.0.2
uvicorn==0.12.3

